Add option to keep [SetInVar] in context after first executation path

+++ Operators/Types/lib/exec/context/SetIntVariable.cs
@@ -1,11 +1,9 @@
 using System;
-using T3.Core;
 using T3.Core.DataTypes;
 using T3.Core.Logging;
 using T3.Core.Operator;
 using T3.Core.Operator.Attributes;
 using T3.Core.Operator.Slots;
-using T3.Core.Resource;
 using T3.Core.Utils;
 
 namespace T3.Operators.Types.Id_7953f704_ebee_498b_8bdd_a2c201dfe278
@@ -19,11 +17,7 @@ namespace T3.Operators.Types.Id_7953f704_ebee_498b_8bdd_a2c201dfe278
         {
             Output.UpdateAction = Update;
         }
-
-        // private T GetEnumFromInput<T>(InputSlot<int> slot, EvaluationContext context) where T:Enum 
-        // {
-        //     return (T)slot.GetValue(context).Clamp(0, Enum.GetValues(typeof(T)).Length - 1);
-        // }
+        
         
         private int ClampForEnum<T>(int i) where T:Enum 
         {
@@ -34,6 +28,7 @@ namespace T3.Operators.Types.Id_7953f704_ebee_498b_8bdd_a2c201dfe278
         {
             var name = VariableName.GetValue(context);
             var newValue = Value.GetValue(context);
+            var clearAfterExecution = ClearAfterExecution.GetValue(context);
 
             var logLevel =  LogLevel.GetEnumValue<LogLevels>(context);
             
@@ -64,7 +59,8 @@ namespace T3.Operators.Types.Id_7953f704_ebee_498b_8bdd_a2c201dfe278
                     if ((int)logLevel >= (int)LogLevels.AllUpdates)
                         Log.Debug($"Setting {name} to {newValue}", this);
 
-                    context.IntVariables.Remove(name);
+                    if(!clearAfterExecution)
+                        context.IntVariables.Remove(name);
                 }
             }
             else
@@ -86,6 +82,8 @@ namespace T3.Operators.Types.Id_7953f704_ebee_498b_8bdd_a2c201dfe278
         [Input(Guid = "4AB2A742-7F3F-4D96-B67E-73E14B4A8F47", MappedType = typeof(LogLevels))]
         public readonly InputSlot<int> LogLevel = new();
 
+        [Input(Guid = "DA431996-4C4C-4CDC-9723-9116BBB5440C")]
+        public readonly InputSlot<bool> ClearAfterExecution = new ();
 
         private enum LogLevels
         {
